zod vlalidations on server

+++ src/actions/actions.ts
@@ -6,9 +6,9 @@ import { revalidatePath } from "next/cache";
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { petFormSchema } from "@/lib/validations";
+import { petFormSchema, petIdeaSchema } from "@/lib/validations";
 
-export async function addPet(newPetData: PetEssentials) {
+export async function addPet(newPetData: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(newPetData);
@@ -28,15 +28,25 @@ export async function addPet(newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], updatedPetData: PetEssentials) {
+export async function editPet(petId: unknown, updatedPetData: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(updatedPetData);
+  if (!validatedPet.success) {
+    return { message: "Invalid data" };
+  }
+
+  const validatedPetId = petIdeaSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { message: "Invalid data" };
+  }
 
   try {
     const updatedPet = await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: updatedPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Could not edit pet" };
@@ -44,13 +54,18 @@ export async function editPet(petId: Pet["id"], updatedPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdeaSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return { message: "Invalid data" };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
